feat(energy): migrate energy API + buffer from int to long
Energy stored/capacity and the four EnergyHandler methods widen to long so
large reactors and battery banks can exceed the ~2.1B int ceiling without
overflow (design doc 7 / 16.1). Pipe-tier throughput stays int (a >2.1B/push
cap is not meaningful). Full suite green (94 tests).

+++ src/main/java/com/chonbosmods/chemistry/api/energy/EnergyHandler.java
@@ -3,17 +3,19 @@ package com.chonbosmods.chemistry.api.energy;
 /**
  * The published energy-capability contract (the Hytale FE-equivalent). Any block may implement
  * this in a few lines; cables and machines transfer against it. Lossless, push-based (design §7).
- * Amounts are non-negative integers in the mod's own energy unit; "simulate" performs no mutation.
+ * Amounts are non-negative {@code long}s in the mod's own energy unit; "simulate" performs no
+ * mutation. Energy is {@code long} (not {@code int}) so large reactors and battery banks can exceed
+ * the ~2.1B {@code int} ceiling without overflow (design §7 / §16.1).
  */
 public interface EnergyHandler {
 
     /** @return amount actually accepted (≤ amount, ≤ free space). */
-    int receiveEnergy(int amount, boolean simulate);
+    long receiveEnergy(long amount, boolean simulate);
 
     /** @return amount actually provided (≤ amount, ≤ stored). */
-    int extractEnergy(int amount, boolean simulate);
+    long extractEnergy(long amount, boolean simulate);
 
-    int getStored();
+    long getStored();
 
-    int getMaxStored();
+    long getMaxStored();
 }
